+++ src/pages/api/v1/company/[companyId]/voucher/index.repository.test.ts
@@ -1,0 +1,15 @@
+// ToDo: (20240605 - Murky) Write test in next pr
+it("should return true", () => {
+  expect(true).toBe(true);
+});
+// import * as module from '@/pages/api/v1/company/[companyId]/voucher/index.repository';
+// import prisma from '@/client';
+// import * as common from '@/lib/utils/common';
+
+// jest.mock('../../../../../../lib/utils/common', () => ({
+//     timestampInSeconds: jest.fn(),
+// }));
+
+// describe('/voucher/index.repository', () => {
+
+// });
